requirements.txt
pywebio
fastapi
uvicorn
amazon-braket-sdk
boto3
Pillow
pytest